requirements.txt
flask
openai==0.28
python-dotenv
pdfplumber
pytesseract
pdf2image